WIP: Fix blitting framebuffers with out-of-range dimensions on NVIDIA GL
NVIDIA GL drivers have an issue where blitFramebuffer fails if the
computed dimensions are outside the valid 32-bit integer range. We now
have a workaround for handling them as a special case in the GL
backend.

The workaround can take two different paths. In case the source and
destination dimensions match, the blitted area can be simply clipped
so that it lies entirely in the positive integer range.

In case the source and destination dimensions don't match, the shader-
based blit is used instead of calling glBlitFramebuffer. The shader-
based blit is updated so that it can clip the destination area to the
destination framebuffer and transform the source area with the same
transform. This is done in case the dimensions are too large to be
used as viewport dimensions.

Transforming the source area needs to be done using double precision
floats for the intermediate values. This way they can accurately
represent all 32-bit integers and numerical errors can't lead to
incorrect blit boundaries. The included end2end tests for the
workaround would fail in case only single precision were to be used.

The shader-based GL blit is is also fixed to handle dimensions
greater than max 32-bit int correctly. To do this, the Rectangle type
is updated to store width and height as 64-bit integers instead of
32-bit.

BUG=chromium:830046
TEST=angle_end2end_tests

+++ src/libANGLE/Context.cpp
@@ -3307,8 +3307,10 @@ void Context::blitFramebuffer(GLint srcX0,
     Framebuffer *drawFramebuffer = mGLState.getDrawFramebuffer();
     ASSERT(drawFramebuffer);
 
-    Rectangle srcArea(srcX0, srcY0, srcX1 - srcX0, srcY1 - srcY0);
-    Rectangle dstArea(dstX0, dstY0, dstX1 - dstX0, dstY1 - dstY0);
+    Rectangle srcArea(srcX0, srcY0, static_cast<int64_t>(srcX1) - srcX0,
+                      static_cast<int64_t>(srcY1) - srcY0);
+    Rectangle dstArea(dstX0, dstY0, static_cast<int64_t>(dstX1) - dstX0,
+                      static_cast<int64_t>(dstY1) - dstY0);
 
     ANGLE_CONTEXT_TRY(syncStateForBlit());
